client: fix signed/unsigned comparision, add -Wextra flag
Among other things, this fixes handling -1 returned from read/write
syscalls, which includes read_watch() being interrupted by a signal.

Fixes QubesOS/qubes-issues#2309

+++ client/qdb-client.c
@@ -202,7 +202,7 @@ static int send_command_to_daemon(qdb_handle_t h, struct qdb_hdr *hdr,
     }
     /* This function writes at most QDB_MAX_DATA bytes (3k) at once,
      * which is atomic on Linux */
-    if (write(h->fd, hdr, sizeof(*hdr)) < sizeof(*hdr)) {
+    if (write(h->fd, hdr, sizeof(*hdr)) < (int)sizeof(*hdr)) {
         /* some fatal error on previous command (and daemon closed connection)
          * perhaps? or daemon has restarted */
         if (errno == EPIPE) {
@@ -216,7 +216,7 @@ static int send_command_to_daemon(qdb_handle_t h, struct qdb_hdr *hdr,
                 return 0;
             } else {
                 /* try again */
-                if (write(h->fd, hdr, sizeof(*hdr)) < sizeof(*hdr))
+                if (write(h->fd, hdr, sizeof(*hdr)) < (int)sizeof(*hdr))
                     return 0;
                 else
                     return 1;
@@ -318,7 +318,7 @@ static int get_response(qdb_handle_t h, struct qdb_hdr *hdr) {
             }
             return 0;
         }
-        if (len < sizeof(*hdr)) {
+        if (len < (int)sizeof(*hdr)) {
             /* partial read?! */
             return 0;
         }
@@ -743,7 +743,7 @@ char *qdb_read_watch(qdb_handle_t h) {
         len = sizeof(hdr);
         if (!QioReadBuffer(h->read_pipe, &hdr, sizeof(hdr))) {
 #else
-        if ((len=read(h->fd, &hdr, sizeof(hdr))) < sizeof(hdr)) {
+        if ((len=read(h->fd, &hdr, sizeof(hdr))) < (int)sizeof(hdr)) {
 #endif
             if (len==0)
                 errno = EPIPE;
